services/dataset-builder/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0
boto3==1.34.110
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
scikit-learn==1.5.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==6.1.0
# test
pytest==8.2.0
pytest-cov==5.0.0
httpx==0.27.0
